+++ CHANGES.rst
@@ -8,6 +8,7 @@ Unreleased
   stopped gracefully.
 * Change nginx ``proxy-body-size`` annotation value as it has stricter validations now
 * Bump ``sql_exporter`` to ``0.17.0`` and fix ``cluster_last_user_activity`` NULL warning.
+* Switch to ``awk`` for zone and node name extraction.
 
 2.43.1 (2025-01-08)
 -------------------